login and registration page move from component diractory to views directory
