+++ src/views/RegistrationPage.vue
